# expression: implement vectorized evaluation for builtinLastDaySig