--- a/expression/builtin_time_vec_test.go
+++ b/expression/builtin_time_vec_test.go
@@ -224,6 +224,9 @@ var vecBuiltinTimeCases = map[string][]vecExprBenchCase{
 		// We test the builtinSysDateWithoutFspSig in TestSysDate function.
 		// {retEvalType: types.ETDatetime},
 	},
+	ast.LastDay: {
+		{retEvalType: types.ETDatetime, childrenTypes: []types.EvalType{types.ETDatetime}},
+	},
 }
 
 func (s *testEvaluatorSuite) TestVectorizedBuiltinTimeEvalOneVec(c *C) {